. ActiveCancelcli.py, ActiveCancelsvr.py, actcancel/maincli.py: added python testcase actcancel.
. rpcif.cpp:
    . fixed a bug in CInterfaceServer's DoInvoke, that previously, when error
    occurs, the response pointer attribute was set to a CConfigDb with
    propReturnValue only, which overwrote the response pointer set by the
    subclass code, and lost some information that the subclass wants to deliver
    to the proxy. Now DoInvoke test and set this pointer only if it is not
    set.

    . fixed a bug in CInterfaceProxy's FillRespData, similiar to the above bug,
    that the old code, when saw the error code in the dbus message, set a
    CConfigDb with propReturnValue only and ignore the CConfigDb along with
    the server message. Now we proceed to extract the CConfigDb from the server
    message and pass it to the CIfIoReqTask as the response data if it exists.
    For ridl serialization, this CConfigDb contains a propSeriProto attribute
    from server, which is important for calling the proxy side callback.

. proxy.sip, proxy.i:
    . added code to extract the propSeriProto attribute from the response data,
    even if the request is failed by server. so that we can pass the correct
    seriProto value to python's HandleAsyncResp, or java's handleAsyncResp.

. genpy.cpp:
    . fixed a bug in proxy side complete callback wrapper, that when error
    returned, the callback was not called and returned. The fix is to generate
    code to pass the error to the callback when error occurs.

. proxy.py: removed numpy.unint32 conversion in AddTimer because exception is
thrown if the timeout second is not an integer but a numpy.uint32. the reason is
unknown yet.

. control.in: fixed a bug in debian package building for Raspberry Pi, that the
libcppunit-1.14-0 and libcppunit-1.15-0 should both be listed, otherwise either
Raspberry Pi or desktop cannot build.

+++ python/rpcf/proxy.py
@@ -281,8 +281,7 @@ class PyRpcServices :
     def AddTimer( self, timeoutSec,
         callback, context )->Tuple[int,cpp.ObjPtr] :
         return self.oInst.AddTimer(
-            np.uint32( timeoutSec ),
-            callback, context )
+            timeoutSec, callback, context )
 
     '''Remove a previously scheduled timer.
     '''
